docs: added some comments, created README as well

+++ expenses/views.py
@@ -1,29 +1,39 @@
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import redirect, render
+from django.db.models.aggregates import Sum
+from django.shortcuts import get_object_or_404, redirect, render
 
 from .forms import ExpensesForm
 from .models import Expenses
 
+"""
+Views module for expenses app.
+All views require a logged in user barring the `index`
+"""
+
 
 # Create your views here.
 def index(request):
+    """Home page view"""
     return render(request, "expenses/index.html")
 
 
 @login_required
 def expenses(request):
+    """View for displaying all expenses of a user"""
     expenses = Expenses.objects.filter(owner=request.user)
-    total_expense = 0
-    for expense in expenses:
-        total_expense += expense.amount
+    total_expenses = expenses.aggregate(total_expense=Sum("amount"))
 
-    context = {"expenses": expenses, "total_expense": total_expense}
+    if total_expenses["total_expense"] is None:
+        total_expenses["total_expense"] = 0
 
+    context = {"expenses": expenses}
+    context.update(total_expenses)
     return render(request, "expenses/expenses.html", context=context)
 
 
 @login_required
 def new_expense(request):
+    """View for adding a new expense"""
     if request.method == "POST":
         form = ExpensesForm(request.POST)
 
@@ -41,7 +51,8 @@ def new_expense(request):
 
 @login_required
 def expense(request, expense_id):
-    expense = Expenses.objects.get(id=expense_id)
+    """View for showing the details of an expense"""
+    expense = get_object_or_404(Expenses, id=expense_id, owner=request.user)
     context = {"expense": expense}
 
     return render(request, template_name="expenses/expense.html", context=context)
@@ -49,19 +60,27 @@ def expense(request, expense_id):
 
 @login_required
 def delete_expense(request, expense_id):
-    expense = Expenses.objects.get(id=expense_id)
+    """
+    View for deleting an expense.
+    Redirects to a template that uses a POST request for deletion.
+    """
+    expense = get_object_or_404(Expenses, id=expense_id, owner=request.user)
 
-    expense.delete()
+    if request.method == "POST":
+        expense.delete()
+        return redirect("expenses:expenses")
 
     return render(request, template_name="expenses/delete_expense.html")
 
 
 @login_required
 def update_expense(request, expense_id):
+    """View for updating an expense"""
     context = {}
-    expense = Expenses.objects.get(id=expense_id)
 
+    expense = Expenses.objects.get(id=expense_id, owner=request.user)
     form = ExpensesForm(request.POST or None, instance=expense)
+
     if form.is_valid():
         form.save()
         return redirect("expenses:expenses")
